reuel/config + example: add functions to add to set / apend to list
The config tree may contain lists, in this case the append() method
can be used instead of set(). Sets are also supported, in this case
the add_to_set() method can be used.

All of these three mentioned functions create the necessary types in
the tree (dict - hash, set, list).